+++ Mond.VSCode/src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import { activateMondDebug } from './activateMondDebug';
 import { activateCompletionProvider } from "./activateCompletionProvider";
 import { activateLanguageFeatures } from "./activateLanguageFeatures";
+import { activateDebugFeatures } from "./activateDebugFeatures";
 import { activateDefinitions } from "./definitions";
 
 export function activate(context: vscode.ExtensionContext) {
@@ -9,6 +10,7 @@ export function activate(context: vscode.ExtensionContext) {
 	activateMondDebug(context);
 	activateCompletionProvider(context);
 	activateLanguageFeatures(context);
+	activateDebugFeatures(context);
 }
 
 export function deactivate() {